requirements.txt
# Core dependencies used across multiple tools
requests>=2.31.0
python-dotenv>=1.0.0

# Optional dependencies for specific tools
# Uncomment as needed:
# beautifulsoup4>=4.12.0  # For HTML parsing in maintenance tools
# schedule>=1.2.0         # For scheduled tasks in monitoring
# click>=8.0.0           # For CLI interfaces
